refactor(session): extract persistent-session management into its own module
Move the persistent-session cluster (start / adopt / stop /
attach / detach / list, plus get_remote_session_buffer) out of the
2000-line SessionManager into a new session::persistent_controller
module, mirroring the file-ops (#2076) and monitoring (#2110) seams.
This is the last and highest-coupling of the three SessionManager
seams (#2111).

PersistentController is a stateless borrowing facade. Unlike FileOps
and MonitoringController, which borrow isolated maps, it borrows the
whole SessionManager: the persistent cluster genuinely drives the
manager's own session lifecycle — start calls create_connection, stop
calls close_session, and attach reuses ensure_output_buffer and
run_output_reader and re-inserts a SessionEntry into the shared
sessions map on the reconnect-after-agent-drop path. The persistent
registry (persistent_sessions) stays a field on the manager; the public
SessionManager methods remain as thin delegators, so the public API,
behavior, error messages, lock ordering, and emitted
persistent-session-state-changed events are unchanged.

Logic is moved verbatim. Manager internals used by the facade
(sessions / agent_manager / persistent_sessions / output_buffers /
session_loggers fields, ensure_output_buffer / run_output_reader
methods, PersistentRecord and the relevant SessionEntry fields) are
widened to pub(super), matching how SessionEntry was already shared
with the file-ops facade.

All 46 session::manager unit tests pass unchanged, including every
persistent-session test and the reconnect / local-skip /
record-preservation regressions. cargo clippy -p termihub
--all-targets is clean.
